[MIG] website_sale_checkout_country_vat: Migration to version 18.0
TT54387
