+++ website_sale_checkout_country_vat/readme/CONFIGURE.md
@@ -2,8 +2,3 @@ The default country that will appear is the one of the current user, or
 the one of the Public User assigned to the website, or the one of the
 company that owns the website. Change it in the corresponding place to
 get a sane default.
-
-If you want to have a sane default also in the country address, check
-the
-[website_sale_default_country](https://github.com/OCA/e-commerce/tree/10.0/website_sale_default_country)
-module.
