+++ website_sale_checkout_country_vat/readme/CONTRIBUTORS.md
@@ -6,3 +6,4 @@
   - Alexandre Díaz
   - Carlos Roca
   - Ernesto Tejeda
+  - Pilar Vargas
