+++ website_sale_checkout_country_vat/static/description/index.html
@@ -392,10 +392,6 @@ autocomplete VAT country code in checkout process.</p>
 the one of the Public User assigned to the website, or the one of the
 company that owns the website. Change it in the corresponding place to
 get a sane default.</p>
-<p>If you want to have a sane default also in the country address, check
-the
-<a class="reference external" href="https://github.com/OCA/e-commerce/tree/10.0/website_sale_default_country">website_sale_default_country</a>
-module.</p>
 </div>
 <div class="section" id="usage">
 <h1><a class="toc-backref" href="#toc-entry-2">Usage</a></h1>
@@ -437,6 +433,7 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li>Alexandre Díaz</li>
 <li>Carlos Roca</li>
 <li>Ernesto Tejeda</li>
+<li>Pilar Vargas</li>
 </ul>
 </li>
 </ul>
