+++ website_sale_checkout_country_vat/views/templates.xml
@@ -10,26 +10,16 @@
         </xpath>
         <xpath expr="//input[@name='vat']" position="after">
             <t t-call="website_snippet_country_dropdown.country_dropdown">
-                <t t-set="complete_field" t-value="'vat'" />
-                <t
-                    t-set="default_value"
-                    t-value="'vat' in checkout and checkout['vat'] or ''"
-                />
-                <t
-                    t-set="country_id"
-                    t-value="'country_id' in checkout and checkout['country_id'] or False"
-                />
+                <t t-set="no_country_field" t-value="'vat'" />
+                <t t-set="default_value" t-value="partner_sudo.vat or ''" />
+                <t t-set="country_id" t-value="country_sudo or False" />
                 <t
                     t-set="default_country"
                     t-value="countries.search(['|', ('code', '=', default_value[:2]), ('id', '=', country_id and int(country_id) or False)], limit=1)"
                 />
                 <t
-                    t-set="no_country_field_extra_classes"
-                    t-value="error.get('vat') and 'is-invalid' or ''"
-                />
-                <t
                     t-set="readonly"
-                    t-value="'1' if 'vat' in checkout and checkout['vat'] and not can_edit_vat else None"
+                    t-value="True if partner_sudo.vat and not can_edit_vat else None"
                 />
             </t>
         </xpath>
